Задание третьего блока дипломной работы

+++ src/main/java/ru/skypro/homework/exception/CommentNotFoundException.java
@@ -1,7 +1,7 @@
 package ru.skypro.homework.exception;
 
 public class CommentNotFoundException extends RuntimeException {
-    public CommentNotFoundException(Integer id) {
+    public CommentNotFoundException(Long id) {
         super("CommentDTO with id " + id + " not found");
     }
 }
